Update modlists.json
## 2.3.6: Vampiric Power
This release swaps out Curse of the Vampire for Sacrilege. Due to this, it is not savefile safe. I would recommend starting a new save or not updating your install if you intend to keep the same install. This change should make vampire playthroughs more satisfying. 

Additionally, this update fixes the crashes related to the Ancient Frost Atronach. This allows for proper completion of Dawnguard. Thank you to Rixban for his help with solving this issue. 

### Mods Added
- Sacrilege - Minimalistic Vampires of Skyrim 2.1.0

### Mods Removed
- Curse of the Vampire

### Patches Updated
- RGE General Patch